Fixes to drawing multiple lines
